+++ packages/cli/test/dev/fixtures/services-v2-bindings/go-api/main.go
@@ -1,0 +1,18 @@
+package main
+
+import (
+	"fmt"
+	"net/http"
+	"os"
+)
+
+func main() {
+	port := os.Getenv("PORT")
+	if port == "" {
+		port = "3000"
+	}
+	http.HandleFunc("/ping", func(w http.ResponseWriter, r *http.Request) {
+		fmt.Fprint(w, "go_api: pong")
+	})
+	http.ListenAndServe(":"+port, nil)
+}